refactor(prompt-area): extract cursor-helpers from use-prompt-area
Moves the 9 pure DOM cursor/selection utilities (saveCursorPosition,
restoreCursorPosition, getCursorOffset, setCursorAtOffset,
createRangeAtOffset, getSelectionOffsets, setSelectionAtOffsets,
findDOMPosition, getTextLengthInRange) from use-prompt-area.ts
(1381 lines → 1166 lines) into a dedicated cursor-helpers.ts module.

The extracted functions take HTMLElement + offsets as inputs and return
Ranges/positions — plain module, no hook wrapper. Invariants documented
in the module header: stay synchronous, never cache Selection/Range
across calls, treat chip nodes atomically via isChipElement.

Bumps registry entry + registry.test.ts to 18 files. All 593 tests pass.
